all styles added
