+++ src/App.css
@@ -48,7 +48,7 @@ nav a {
 
 .input-wrapper {
   margin-bottom: 1rem;
-  color: #6c757d;
+  color: #080808;
   display: block;
   font-weight: bold;
 }
@@ -74,7 +74,9 @@ button {
 }
 
 button:disabled {
-  background-color: #6c757d;
+  background-color: #747272;
+  height: 50px;
+  width: 100%;
 }
 .Home-card {
   max-width: 1200px;
@@ -165,7 +167,7 @@ button:disabled {
 }
 
 .room-detail img {
-  width: 40%;
+  width: 100%;
   border-radius: 8px;
   margin-bottom: 20px;
 }
@@ -224,7 +226,8 @@ button:disabled {
 }
 
 .reservation-img {
-  width: 50%;
+  width: 90%;
+  height: 90%;
 }
 
 .row {
@@ -233,6 +236,7 @@ button:disabled {
 
 .col {
   width: 100%;
+  text-align: center;
 }
 .unavailable {
   width: 300px;
@@ -248,7 +252,7 @@ button:disabled {
   box-shadow: 0 0 13px rgba(19, 19, 19, 0.1);
 }
 .room-data {
-  max-width: 40%;
+  max-width: 50%;
   padding: 1.5rem;
   background-color: #fff;
   border-radius: 10px;
@@ -256,9 +260,10 @@ button:disabled {
   margin: 0 auto;
 }
 .container2 {
-  max-width: 40%;
-
-  padding: 10px;
+  max-width: 51%;
+  margin: 0 auto;
+  margin-top: 2%;
+  padding: 20px;
   background-color: #fff;
   border-radius: 10px;
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.671);
@@ -267,7 +272,7 @@ button:disabled {
   margin-top: 10px;
 }
 .checkout {
-  margin-left: 20px;
+  margin-left: 50px;
   margin-top: 20px;
   margin-bottom: 20px;
   font-weight: bold;
@@ -286,4 +291,114 @@ button:disabled {
   background-color: #fff;
   border-radius: 10px;
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.671);
+}
+.yes {
+  margin-bottom: 10px;
+}
+.buttons {
+  margin: 2px;
+}
+.edit-bold {
+  font-weight: bold;
+}
+.container3 {
+  max-width: 800px;
+  width: 100%;
+  padding: 2rem;
+  background-color: #fff;
+  border-radius: 10px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.671);
+  margin: 0 auto;
+  margin-top: 6%;
+}
+.register {
+  text-align: center;
+}
+.centered {
+  align-items: center;
+}
+
+.reservation-card {
+  display: flex;
+  flex-direction: column;
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  margin: 10px 0;
+  padding: 20px;
+  background-color: #f9f9f9;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.reservation-content {
+  display: flex;
+  align-items: center;
+}
+
+.reservation-img-container {
+  height: 150px;
+  margin-right: 20px;
+  flex: 0 0 150px;
+}
+
+.reservation-img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border-radius: 8px;
+}
+
+.reservation-details {
+  flex: 1;
+  padding: 0 20px;
+}
+
+.check-in,
+.check-out {
+  margin: 10px 0;
+}
+
+.reservation-buttons {
+  width: 150px;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.reservation-buttons button {
+  padding: 10px 15px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 16px;
+}
+
+.btn-edit {
+  background-color: #595f09;
+  color: #fff;
+}
+
+.btn-delete {
+  background-color: #681106;
+  color: #fff;
+}
+
+.btn-review {
+  background-color: #065168;
+  color: #fff;
+}
+.cenr {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+}
+.container4 {
+  text-align: center;
+  max-width: 500px;
+  width: 100%;
+  padding: 1.5rem;
+  background-color: #fff;
+  border-radius: 10px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.671);
+  margin: 0 auto;
+  margin-top: 7%;
 }
