+++ src/pages/EditReservForm.jsx
@@ -26,15 +26,15 @@ const EditReservForm = () => {
         newCheckOut
       })
 
-      const response = await Client.put(
-        `/reservations/${reservationId}`,
-        { oldCheckIn, oldCheckOut, newCheckIn, newCheckOut }
-      )
+      const response = await Client.put(`/reservations/${reservationId}`, {
+        oldCheckIn,
+        oldCheckOut,
+        newCheckIn,
+        newCheckOut
+      })
 
-      alert('Reservation updated successfully!')
       navigate('/')
     } catch (error) {
-
       console.error(
         'Error updating reservation:',
         error.response ? error.response.data : error.message
@@ -51,39 +51,49 @@ const EditReservForm = () => {
   }
 
   return (
-    <div>
+    <div className="container">
       <h2>Edit Reservation</h2>
-   
-      <p>Old Check-In Date: {new Date(oldCheckIn).toLocaleString()}</p>
-      <p>Old Check-Out Date: {new Date(oldCheckOut).toLocaleString()}</p>
+
+      <p className="edit-bold">
+        Old Check-In Date: {new Date(oldCheckIn).toLocaleString()}
+      </p>
+      <p className="edit-bold">
+        Old Check-Out Date: {new Date(oldCheckOut).toLocaleString()}
+      </p>
       {confirmEdit ? (
         <>
-      
           <div>
-            <label>New Check-In Date and Time:</label>
-            <input
-              type="datetime-local"
-              value={newCheckIn}
-              onChange={(e) => setNewCheckIn(e.target.value)}
-            />
+            <p className="edit-bold">
+              <label>New Check-In Date and Time:</label>
+              <input
+                type="datetime-local"
+                value={newCheckIn}
+                onChange={(e) => setNewCheckIn(e.target.value)}
+              />
+            </p>
           </div>
           <div>
-            <label>New Check-Out Date and Time:</label>
-            <input
-              type="datetime-local"
-              value={newCheckOut}
-              onChange={(e) => setNewCheckOut(e.target.value)}
-            />
+            <p className="edit-bold">
+              <label>New Check-Out Date and Time:</label>
+              <input
+                type="datetime-local"
+                value={newCheckOut}
+                onChange={(e) => setNewCheckOut(e.target.value)}
+              />
+            </p>
           </div>
-       
+
           <button onClick={handleUpdate}>Update Reservation</button>
         </>
       ) : (
         <>
-         
           <p>Are you sure you want to change the reservation dates?</p>
-          <button onClick={handleConfirm}>Yes</button>
-          <button onClick={() => navigate(-1)}>No</button>
+          <button className="yes" onClick={handleConfirm}>
+            Yes
+          </button>
+          <button className="no" onClick={() => navigate(-1)}>
+            No
+          </button>
         </>
       )}
     </div>
