+++ src/pages/Register.jsx
@@ -1,0 +1,102 @@
+import '../App.css'
+
+import { useState } from 'react'
+import { useNavigate } from 'react-router-dom'
+import { RegisterUser } from '../services/Auth'
+
+const Register = () => {
+  let navigate = useNavigate()
+
+  const [formValues, setFormValues] = useState({
+    name: '',
+    email: '',
+    password: '',
+    confirmPassword: ''
+  })
+
+  const handleChange = (e) => {
+    setFormValues({ ...formValues, [e.target.name]: e.target.value })
+  }
+
+  const handleSubmit = async (e) => {
+    e.preventDefault()
+    await RegisterUser({
+      name: formValues.name,
+      email: formValues.email,
+      password: formValues.password
+    })
+    setFormValues({
+      name: '',
+      email: '',
+      password: '',
+      confirmPassword: ''
+    })
+    navigate('/signin')
+  }
+
+  return (
+    <>
+      <div className="container3">
+        <h1 className="register">Register</h1>
+        <div className="register-card centered">
+          <form onSubmit={handleSubmit} className="reg">
+            <div className="input-wrapper">
+              <label htmlFor="name">Name</label>
+              <input
+                onChange={handleChange}
+                name="name"
+                type="text"
+                placeholder="John Smith"
+                value={formValues.name}
+                required
+              />
+            </div>
+            <div className="input-wrapper">
+              <label htmlFor="email">Email</label>
+              <input
+                onChange={handleChange}
+                name="email"
+                type="email"
+                placeholder="example@example.com"
+                value={formValues.email}
+                required
+              />
+            </div>
+
+            <div className="input-wrapper">
+              <label htmlFor="password">Password</label>
+              <input
+                onChange={handleChange}
+                type="password"
+                name="password"
+                value={formValues.password}
+                required
+              />
+            </div>
+            <div className="input-wrapper">
+              <label htmlFor="confirmPassword">Confirm Password</label>
+              <input
+                onChange={handleChange}
+                type="password"
+                name="confirmPassword"
+                value={formValues.confirmPassword}
+                required
+              />
+            </div>
+            <button
+              disabled={
+                !formValues.email ||
+                (!formValues.password &&
+                  formValues.confirmPassword === formValues.password)
+              }
+            >
+              Register
+            </button>
+          </form>
+        </div>
+      </div>
+    </>
+  )
+}
+
+export default Register
