+++ src/pages/SignIn.jsx
@@ -6,7 +6,7 @@ import { SignInUser } from '../services/Auth'
 
 const SignIn = ({ setUser }) => {
   let navigate = useNavigate()
-  
+
   const [formValues, setFormValues] = useState({ email: '', password: '' })
 
   const handleChange = (e) => {
@@ -22,7 +22,8 @@ const SignIn = ({ setUser }) => {
   }
 
   return (
-    <div className="container wrapper">
+    <div className="container3">
+      <h1 className="register">Sign In</h1>
       <div className="signin-card">
         <form onSubmit={handleSubmit}>
           <div className="input-wrapper">
@@ -51,8 +52,7 @@ const SignIn = ({ setUser }) => {
           </button>
         </form>
       </div>
-      </div>
-
+    </div>
   )
 }
